0308 第二章音樂音效增加、成就黑白鎖更正、第二關關卡圖更正、Loading頁面修正OK、畫符點collider修正OK、關卡資訊回復OK

+++ Assets/Sprite/StaticObject.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public static class StaticObject
 {
-	public static int nowClass = 0;
+	public static float nowClass = 0;
 	//------------------音量音效---------------------
 	public static float bgmVolume = 1f;
 	public static float sfxVolume = 1f;
